#3 update module names
- the autoloader in rails expects a specific naming of modules and classes
 based on the directory structure
  - since Rails 6 zeitwerk is used as autoloader, Rails 7 does not support the
   classic autoloader at all
